Bring back JVM args to non-Spark kernels.
