+++ polynote-kernel/src/main/scala/polynote/kernel/remote/DeploySparkSubmit.scala
@@ -46,23 +46,12 @@ object DeploySparkSubmit extends DeployCommand {
       sparkConfig.get("sparkSubmitArgs").toList.flatMap(parseQuotedArgs)
 
     val isRemote = sparkConfig.get("spark.submit.deployMode") contains "cluster"
-    val libraryPath = List(sys.props.get("java.library.path"), sys.env.get("LD_LIBRARY_PATH"))
-      .flatten
-      .map(_.trim().stripPrefix(File.pathSeparator).stripSuffix(File.pathSeparator))
-      .mkString(File.pathSeparator)
-
-    val javaOptions = Map(
-      "log4j.configuration" -> "log4j.properties",
-      "java.library.path"   -> libraryPath
-    )
 
     val allDriverOptions = {
       nbConfig.jvmArgs.toList ++
       sparkConfig.get("spark.driver.extraJavaOptions").toList ++
-      javaOptions.toList.map {
-        case (name, value) => s"-D$name=$value"
-      } mkString " "
-    }
+      asPropString(javaOptions)
+    } mkString " "
 
     val additionalJars = classPath.toList.filter(_.getFile.endsWith(".jar"))
